requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
html2text>=2020.1.16
PyYAML>=6.0

# RAG Chatbot Dependencies
streamlit>=1.28.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
openai>=1.0.0
torch>=2.0.0
transformers>=4.30.0 